reorganized files suitable for python package
